+++ src/Widgets/TestCases.hpp
@@ -79,6 +79,11 @@ class TestCases : public QWidget
 
     void setTestCaseEditFont(const QFont &font);
 
+    void updateHeights();
+
+    QVariantList splitterStates() const;
+    void restoreSplitterStates(const QVariantList &states);
+
   public slots:
     void setVerdict(int index, Core::Checker::Verdict verdict);